refactor: extract header components
