+++ web/components/builder/header/HeaderInspector.tsx
@@ -1,0 +1,94 @@
+'use client'
+import React from 'react'
+import { useBuilderStore, type Elm } from '@/store/builderStore'
+
+export function HeaderInspector({ elm }: { elm: Elm }) {
+  const updateProps = useBuilderStore((s) => s.updateProps)
+  return (
+    <>
+      <div className="text-sm font-medium">Login Button</div>
+      <div className="space-y-2 text-xs">
+        <label className="flex items-center gap-2">
+          <input
+            type="checkbox"
+            checked={elm.props?.loginButton?.enabled ?? false}
+            onChange={(e) => {
+              if (e.target.checked) {
+                const cur = elm.props?.loginButton
+                updateProps(elm.id, {
+                  loginButton: {
+                    enabled: true,
+                    label: cur?.label ?? 'Log in',
+                    variant: cur?.variant ?? 'solid',
+                    href: cur?.href,
+                  },
+                })
+              } else {
+                updateProps(elm.id, {
+                  loginButton: { ...(elm.props?.loginButton ?? {}), enabled: false },
+                })
+              }
+            }}
+          />
+          Enable
+        </label>
+        {elm.props?.loginButton?.enabled && (
+          <>
+            <label className="flex flex-col gap-1">
+              Label
+              <input
+                className="px-2 py-1 rounded bg-zinc-900 border border-zinc-800"
+                value={elm.props?.loginButton?.label ?? ''}
+                onChange={(e) =>
+                  updateProps(elm.id, {
+                    loginButton: {
+                      ...(elm.props?.loginButton ?? {}),
+                      label: e.target.value,
+                    },
+                  })
+                }
+                type="text"
+              />
+            </label>
+            <label className="flex flex-col gap-1">
+              Variant
+              <select
+                className="px-2 py-1 rounded bg-zinc-900 border border-zinc-800"
+                value={elm.props?.loginButton?.variant ?? 'solid'}
+                onChange={(e) =>
+                  updateProps(elm.id, {
+                    loginButton: {
+                      ...(elm.props?.loginButton ?? {}),
+                      variant: e.target.value as 'solid' | 'outline',
+                    },
+                  })
+                }
+              >
+                <option value="solid">solid</option>
+                <option value="outline">outline</option>
+              </select>
+            </label>
+            <label className="flex flex-col gap-1">
+              Link
+              <input
+                className="px-2 py-1 rounded bg-zinc-900 border border-zinc-800"
+                value={elm.props?.loginButton?.href ?? ''}
+                onChange={(e) =>
+                  updateProps(elm.id, {
+                    loginButton: {
+                      ...(elm.props?.loginButton ?? {}),
+                      href: e.target.value || undefined,
+                    },
+                  })
+                }
+                type="text"
+                placeholder="/login"
+              />
+            </label>
+          </>
+        )}
+      </div>
+    </>
+  )
+}
+
